# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Direction of transformation
 */
export type TransformPhotoBodyDirection =
  (typeof TransformPhotoBodyDirection)[keyof typeof TransformPhotoBodyDirection];

export const TransformPhotoBodyDirection = {
  dayToNight: "dayToNight",
  nightToDay: "nightToDay",
} as const;

export interface TransformPhotoBody {
  /** Base64-encoded source image (without data URL prefix) */
  imageBase64: string;
  /** MIME type of the source image (e.g. image/jpeg, image/png) */
  mimeType: string;
  /** Direction of transformation */
  direction: TransformPhotoBodyDirection;
}

export interface TransformPhotoResponse {
  imageBase64: string;
  mimeType: string;
}