# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Transform a photo from day to night or night to day
 */
export const TransformPhotoBody = zod.object({
  imageBase64: zod
    .string()
    .describe("Base64-encoded source image (without data URL prefix)"),
  mimeType: zod
    .string()
    .describe("MIME type of the source image (e.g. image\/jpeg, image\/png)"),
  direction: zod
    .enum(["dayToNight", "nightToDay"])
    .describe("Direction of transformation"),
});

export const TransformPhotoResponse = zod.object({
  imageBase64: zod.string(),
  mimeType: zod.string(),
});