# lib/api-zod/src/generated/types/transformPhotoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransformPhotoBodyDirection } from "./transformPhotoBodyDirection";

export interface TransformPhotoBody {
  /** Base64-encoded source image (without data URL prefix) */
  imageBase64: string;
  /** MIME type of the source image (e.g. image/jpeg, image/png) */
  mimeType: string;
  /** Direction of transformation */
  direction: TransformPhotoBodyDirection;
}